Skip mid-epoch bad-size batches
Bad-size batches trigger graph compiles
Currently, we either take the hit, or if drop_last is enabled, terminate
the epoch early.
This will skip the batch and continue the epoch.

This was needed in context of Fairseq as Fairseq shuffles the data in
batches and the last batch ends up in the middle.