.circleci: Make the grabbing of the docker tag dynamic
upstream docker tags were changed to be based on the output of pytorch's
`git rev-parse HEAD:.circleci/docker` so we should calculate docker tags
here the same way since we want to use the most up to date upstream
image.

This is no guarantee that the image will actually be there, but if this
image is borked then upstream is most likely borked as well so we'll
just need to fix it all around.

Signed-off-by: Eli Uriegas <eliuriegas@fb.com>