Replace Usages of xla:util::Hash* with torch::lazy::Hash*
As part of migration of core lazy tensor functionality to PyTorch core,
this uses the newly added torch::lazy::Hash functions from PyTorch.

Note: while the Hash* functions are largely identical to the original
XLA ones, the underlying uint128 class is from protobuf instead of absl,
since it was a slightly smaller dependency to ingest and get building on
multiple OS/platform combinations for PyTorch.