[Distributed] Companion fix after pytorch PR73164
Author: @esaliya

Summary:
PyTorch PR pytorch/pytorch#73164 adds changes to extend the 3rd party
distributed backend registration. As part of this the internal data
structure is changed to a named tuple of {creator_fn, bool}. This
causes some of the tests in xla to fail as it expects a callable
being returned for a particular backend name.

This PR fixes this issue by calling the creator_fn of the namedtuple
instead of the namedtuple itself.

Test Plan:
python test/test_torch_distributed_xla_backend.py