[Functionalization] Fix SPMD tests
Summary:
This pull request tries to fix three SPMD tests that fails with Functionalization.
1. test_optimizer_step_with_sharding: Currently SPMD relies on tensor.add_ to
populate sharding spec for weights across different iterations. However, Functionalization
removes mutations and therefore this method is not working anymore. To reconcile that,
the SPMD team is trying to figure out a new place to store the sharding spec. Potentially,
the DTensor or XLAShardedTensor such that the python object is the ultimate source of truth and
can be kept intact regardless any dispatcher magic. We shouldn't need propagation in this case.
2. test_inplace_add_with_sharding: See above.
3. test_model_weight_metrics: Before Functionalization, model.weight is an at::tensor stored
in the XLATensor, but it's an IRValue now. Therefore, the test no longer holds true.

Test Plan:
CI.

Fix linters

Address comments

Nit