[Functionalization] Properly skip test_empty_strided (#4609)
Summary:
Functionalization breaks torch.autograd.grad for us. The reason to skip
the test is due to the followings:
1. We don't have other test cases or use cases that suggest we officially
support torch.autograd.grad.
2. This test case suggested that it's testing test_empty_strided and not
torch.autograd.grad.
3. We have cpp test: AtenXlaTensorTest.TestEmptyStrided to test empty_strided
and it passes.
Therefore, we believe temporarily breaking torch.autograd.grad should be
okay and we should skip the test. pytorch/pytorch#91199 is tracking the
progress of the fix.

Test Plan:
CI.