Use Ansible for building wheels and provisioning docker images. (#4531)
The first version of Ansible playbook that can build PyTorch and XLA wheels and provision docker image (release stage).
The playbook can install different pip / apt packages and set environment variable depending on the specified stage, architecture and accelerator.

The playbook we'll be later invoked in a dockerfile, so the config structure or Ansible roles may change a little in the future.

Co-authored-by: Mateusz Lewko <mlewko@google.com>