[FSDP] Add meta device and torchdistx deferred_init feature in FSDP (#4664)
This PR brings the deferred initialization from torch FSDP into torch-xla FSDP

To use the FSDP with auto_wrap, we will likely to encounter a OOM issue when instantiate the gigantic model before wrapping it with FSDP.

model = gigantic_model() ----> OOM here
model = XlaFullyShardedDataParallel(
      model,
      auto_wrap_policy=always_wrap
)
To address this issue, we need a mechanism to delay the initialization to the FSDP stage, where we should initialize the model layer by layer (module by module) and partition the parameters immediately to reduce the peak memory. To achieve this, we need two things: 1) During the instantiation we should not allocate the memory immediately. 2) Re-initialize the sub module correctly.

There are two ways to delay the memory allocation. The first one is the meta device. The actual memory allocation will not happen to the tensors with meta device, and only in the FSDP stage we move the tensors from meta device to xla device, allocate the memory and re-do the initialization.

module = MyModule(device="meta")
def my_init_fn(module):
   # responsible for initializing a module, such as with reset_parameters
   ...
fsdp_model = FSDP(module, param_init_fn=my_init_fn, auto_wrap_policy=size_based_auto_wrap_policy)
However, this method has a duplicated-initalization problem that is also raised in PyTorch-FSDP pytorch/pytorch#90465. This problem will cause the initialization result mismatches.

The second approach is to use torchdistX package. Once the module is wrapped by deferred_init, all the tensors inside will become fake tensors. And It provides extra ability to record the operations applied to the fake tensors. When we want to materialize the module, it will allocate the memory and replay all the operations.

# With torchdistX
module = deferred_init.deferred_init(MyModule)
# Will initialize via deferred_init.materialize_module().
fsdp_model = FSDP(module, auto_wrap_policy=size_based_auto_wrap_policy)
TorchdistX part is tested locally with source-build torchdistx. The disadvantage of this approach is that the official support of torchdistX is till PT1.12. I tried the other versions. It can work with PT-XLA 1.13 out-of-the-box and need minor changes to work with latest PT. I don't know why they stop maintaining this package. Feel free to comment on it if anyone has insights.