Do not fail on lack of default precision set.
I discovered some models from the suite do not have the precision set
so instead of failing the script we just log the case, and use the default
precision, as no additional machinery should run for the Inductor
anyway. Additionally I wrapped the exceptions with the
ValueError so the logging message will not pollute with info about
str not inheriting from Exception class.

ecg@, note that needs to be hooked "somewhere". Not sure where, as
there was a revert in https://github.com/pytorch/xla/pull/6134.