Measure GPU wall time without launch overheads. (#6233)
This is useful for scoping up the single graph, and seeing how compiler is able to optimise it.
This is is optional i.e not added by default. Once we start setting up another benchmark suite with profiling info etc. this should be run additionally, but separately so CUPTI interface won't interact with pure CUDA events.

We do not measure compilation time for this option yet.