aggregate.py: fix detection of no datapoints
The existing check became obsolete once we added the --backends flag
in PR#6392; in particular, len(speedups) could be 1 when a single
backend is passed to --backends. Fix it and add a test to make sure
we emit no output to stdout (note that the warning message goes to
stderr).