Update export.py
Export the names and raw ndarray instead of the Jax array when export_numpy is enabled. The Jax sharding cannot be directly applied to Jax array that's already placed on CPU. You will need to manually call device_put + pmap, which is troublesome and error prone. The better way is to bind the sharding spec together with the ndarray and let Jax handle the sharding for the user.