Correct the cxx11_abi condition
The cxx11_abi is a number variable that is either 0 or 1. Unfortunately,
when specifying it in the `when:` ansible stanza, a numeric value of 0
may get interpreted as true. This causes our nightly wheels from Oct 28
to Nov 04 to be renamed with a "cxx11" suffix, regardless if they
actually used C++11 ABI.