xm.save() should not set sync_xla_data=True when sync'ing.
Setting sync_xla_data=True performs tensor graph sync as if
it's a mark step, which triggers buffer aliasing to be performed.
However, it's not safe to do so unless all live tensors are
being sync'd.

Also fix torch_xla.utils.serialization.save() which has the same
issue.

This fixes #8422