Ability to explicitly shard the gradient for the weights of an einsum/linear layer
When both multiplicands have sharded contraction dimensions, that
sharding annotation will disappear from the result, because we have
summed over the contraction dimension. This leads to suboptimal
sharding decisions later on.

This extra arg lets you specify the sharding again for grad_weights,
which is the result of einsum(grad_outputs, inputs), both of which could
easily be sharded e.g. over the batch dimension which will disapper.