[scan] Test we don't recompile under debugging env flags
When XLA_HLO_DEBUG=1 is set, the lowered HLO contains extra scope and
line number information. When a function is scanned twice, the line
numbers stay the same but the scopes will change. This tests that those
differences don't cause the graph hash of the scan computation to
change, so that we don't recompile on every scan call.