[scan] Handle fn output aliasing an input
When the combine function outputs a tensor unchanged from its input e.g.
it passes the input x as the output y, the graph extracted by
AOTAutograd won't have that pair of variables. That's because
AOTAutograd wants to handle aliasing and input mutation with regular
code as opposed to inside the FX Graph. That's bad news for scan since
we would like a graph with all the inputs/outputs. So we'll add a clone
when needed to defeat this AOTAutograd "optimization".

Also add a test that fn inputs cannot be in-place modified.