update modes
 Distinguish when someone enables env via context manager vs.
  enable_globally.
The reason is that: if a user calls `torch.ones` we want to respect
the device, specially meta. However, if this op is called as part
of a lowering, then we need to replace with jax device.

Currently there is no way for C++ dicomposed ops to know if a tensor
is of device jax