Revamp SPMD guide
I was going to document some device ID assignment tricks and then I
realized in order to explain that we need to first explain what is a
device mesh. The existing documentation is very poor so I'm adding an
in-depth guide in this PR.

While going over the APIs used I noticed the API documentation and
typing also has some problems so I fixed those while I'm here.