Refactor call_jax to allow implementing ops in python
This PR introduces FlattenedInputFunc abstraction,
which is an refactor of jax_func_to_xla_computation
exposing the closures insides of it as methods.

FlattenedInputFunc itself splits a function into
3 components. (preprocess, flat_call, postprocess);
among them, flat_call will take a single flattened
list/tuple as input ans returns a flattened list
as output. The way to think of it is to treat it
as a triplet of functions, that are expected to
be called in this exact order. The inner function
has no restrictions, it can be a jax function,
an function that takes and returns XlaOps or any other.

The above splitting is useful because compiling a
function to xla will have to be a flat function.

Then, I abstracted the cache management to be
something that any FlattenedInputFunc can share,
as long as we know how to "especialize" or create
shape specialized functions.

With this refactor, JaxCallable will replace
previous capability on call_jax, and XlaCallable
can call into Python functions using xla_builder
API directly.