Install libtpu directly instead of torch_xla[tpu]
This is the list of packages that is installed in a dev container. 

When supporting a new python version for torch_xla, trying to install torch_xla[tpu] in the devcontainer creates a circular dependency where torch_xla for that python version already needs to exist. To break the circular dependency, we can directly install the latest libtpu stable version instead of using torch_xla[tpu]