Actually make Python 3.12 work
This PR defaults the dev container to Python 3.12 because Python 3.10 is
soon unusable with future JAX versions.

Python 3.12 builds a package for editing in an isolated pip environment.
That means we need a pyproject.toml file declaring which packages
setup.py depends on, which is added in this PR.

Also setup.py depends on torch to get the C++11 ABI status but that just
won't work in Python 3.12 so that logic is adjusted accordingly. We need
a bunch of other fixes due to the isolated environment change in Python
3.12.

Finally, we need to add `12` into the list of Python versions when
generating JAX dependencies.

TEST: `scripts/build_developer.sh -a -t`