Remove tests that are defined outside of this repo. (#9577)
Includes:
* TPU info
* tests in pytorch/pytorch

With this PR, I'd like to stablish the convention of tests: we test
stuff defined in this repo only. If we want to test an interation of
torch_xla with a third_party library, we would need to define the test
itself inside of this repo.

Reason for disabling TPU info test:
Historically tpu-info commandline tool's source is in this repo, so the
test is in this repo. It's source was then moved to
https://github.com/AI-Hypercomputer/cloud-accelerator-diagnostics
because it's not a Pytorch/XLA specific tool. Now, given that it is
released independently, we cannot control it. (see
https://github.com/pytorch/xla/issues/9568)

Reason for disabling pytorch/pytorch:
The test themselves are defined in pytorch/pytorch instead of
pytorch/xla repo. (the ../.. path goes to the parent folder which is
where pytorch/pytorch is cloned).
If we find a particular test helpful, we should copy that test into
pytorch/xla
Historically this list are a tiny subset of the full pytorch/pytorch
tests, and is often commented out if failing. i.e. this is merely a list
of tests that we found out that happen to pass, not a list that we want
to enforce passing.