Remove CUDA plugin. (#9597)
This PR removes the CUDA plugin in `plugins/cuda` directory, as well as
explicit mentions and usages. This is in line with the CUDA deprecation
that started on release 2.8.

There are still a few mentions to `torch_xla_cuda_plugin` at
`.github/ci.md` markdown file. But, I will leave those for a future PR,
since we would need to re-write those parts.