Increase CI build timeout. (#9694)
PRs from external repositories are timeouting on `_build_torch_xla.yml`
workflow. That's because in those cases, [the remote cache is
disabled][1]. In such cases, [the fixed 45 minutes][2] is not enough
anymore. See, for example, PR #9682 that fails due to this timeout.

Here's my plan to address this issue:

- Bump the timeout by 5 minutes (this PR)
- Create a disk-cache using GitHub cache actions for reducing build time
on PRs from external repositories (see [#9659][3] for more information)

This PR will go through the following steps:

- [x] Reproduce the CI build timeout 
- [x] Bump the timeout by 5 minutes

[1]:
https://github.com/pytorch/xla/blob/df6798dfb931ce7c7fe5bed2447cd1092a5981af/.github/workflows/_build_torch_xla.yml#L36
[2]:
https://github.com/pytorch/xla/blob/df6798dfb931ce7c7fe5bed2447cd1092a5981af/.github/workflows/build_and_test.yml#L44
[3]: https://github.com/pytorch/xla/pull/9659