Improve error handling and error message of device string parsing. (#9710)
This PR improves error handling and error messages of the
`ParseDeviceString` function, which is used for parsing a given device
string into a `BackendDevice`. Besides that, it also applies the
improvement to `GetDefaultDevice` function, which makes use of the
parsing function.

**Key Changes:**

- `ParseDeviceString` is deprecated in favor of `SafeParseDeviceString`
- `GetDefaultDevice` is deprecated in favor of `SafeGetDefaultDevice`
- Inner lambda of `GetDefaultDevice` for initializing the function-local
static variable was refactored into a static function
- New `test_device.cpp` file was created for testing the new
`SafeParseDeviceString` error messages